% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{impact_scaffold_project}
\alias{impact_scaffold_project}
\title{Scaffold an IMPACT Project}
\usage{
impact_scaffold_project(
  project_path = ".",
  use_rstudio = TRUE,
  use_git = TRUE,
  use_targets = TRUE,
  core_packages = default_core_packages,
  ...
)
}
\arguments{
\item{project_path}{A character string. The path to the directory where the
project will be created. Defaults to the current working directory (\code{"."}).
If the directory does not exist, it will be created.}

\item{use_rstudio}{Logical. If \code{TRUE} (the default), an RStudio Project
(\code{.Rproj} file) will be created.}

\item{use_git}{Logical. If \code{TRUE} (the default), a Git repository will be
initialized in the project directory (e.g., using \code{usethis::use_git()}).}

\item{use_targets}{Logical. If \code{TRUE} (the default), a basic {targets}
pipeline will be set up (e.g., by creating \verb{_targets.R} using
\code{targets::use_targets()}).}

\item{core_packages}{A character vector of R packages to install into the
project's {renv} library and include in \code{renv.lock}. This can include
packages from CRAN, Bioconductor, or GitHub (e.g.,
\code{"username/repo"}). Defaults to \code{default_core_packages} which includes
packages like \code{"impact-initiatives/cleaningtools"}, \code{"targets"},
\code{"tarchetypes"}, and \code{"visNetwork"}.}

\item{...}{Additional arguments, potentially to be passed to underlying
functions like \code{renv::install} (via \code{impact_renv_init}). (Currently not
fully implemented for passthrough).}
}
\value{
The \code{project_path} (invisibly), which is the absolute path to the
newly created project.
}
\description{
This function automates the setup of a new project directory. It creates the
directory (if it doesn't exist), initializes an RStudio project, a Git
repository, and a {targets} pipeline structure based on the arguments
provided.
It also initializes an {renv} environment and installs a specified set of
core R packages, creating an \code{renv.lock} file.
}
\details{
Initializes a new standardized project structure for analysis with {renv},
and optionally including setup for RStudio, Git, {targets} environment with
pre-defined core packages.
}
\examples{
\dontrun{
impact_scaffold_project("my_new_analysis")

# Create a project in the current directory, but without Git integration
# and with a custom set of core packages
impact_scaffold_project(
  project_path = ".",
  use_git = FALSE,
  core_packages = c("dplyr", "ggplot2", "mycustomtool@v1.0")
)

# Create a project without targets setup
impact_scaffold_project("project_without_targets", use_targets = FALSE)
}
}
